Izmene.

+++ Registrator/src/main/java/app/entities/Admin.java
@@ -19,7 +19,7 @@ public class Admin {
 		
 	}
 
-	public long getIdAdmina() {
+	public int getIdAdmina() {
 		return idAdmina;
 	}
